SMS(project-root)/requirements.txt
# Production Dependencies
python-telegram-bot[job-queue]>=22.0,<23.0
python-dotenv==1.0.0
aiohttp==3.9.5
tinydb==4.8.0
psutil==5.9.0